camera.py renamed to base_camera.py
