+++ color_tracker/utils/camera/__init__.py
@@ -1,2 +1,2 @@
-from .camera import Camera
+from .base_camera import Camera
 from .web_camera import WebCamera
